Created DataStorageManager Class
This class reads and writes to files passed in by the Activity using
this storage manager.  It has not been tested, it will once I have a
Activity that will use it to read and write files. It has static
methods, so it does not have to be created to be used.