Completed the Initial Concept of Appointments
The user can create new appointments, view existing appointments in a
list, and view the details of each appointment. There is still work
needed to be done. There is no way to delete or edit existing
appointments yet. Also refactoring and commenting is needed for each of
these classes is required.