Update Notifications to fix crash error
Something caused the notifications system to crash the application. The
services used in the application would not be able to be started, and
the application would crash. When the service is started, it will
execute some code, then it will stop itself. With this, the service just
does some work, and then stops. The application would not crash due to
unexpected service destroys or stops....