Created A Notes Feature
The Notes Feature is somewhat complete. It works, a list of notes is
given to the user when selecting the notes feature. There is an option
to add a note as well. When selecting a note, it is displayed to the
user. The user then can select to edit the note or delete the note. This
was developed based on the Appointments feature, with a couple of
differences. There is more that can be done, but this is a start.