Fixed an Issue when the user selects the notification
There was a bug when the user selected a notification message and tries
to delete the appointment associated with it. The user is not given an
option to delete an appointment or edit it when the user clicks on the
notification. This may change when a possible solution is found