Fixed Null Intent Issue with Notifications
There was an error with the notifications intents not being sent
properly. It had to do with again the uniqueness of the intent being
sent. A unique request code was given to the pending intent to aid with
improper updating of intents or losing intents. This really took just
one parameter to fix. There may be other issues, so more testing is
necessary.