Update Notifications to Account for Medications
Notifications has some structure set for Medications, so that it can
generate notifications for medications. Though most of the logic still
needs to be implemented on the medications part of the application. The
notifications set up is specific, instead of generic. There is still
code that needs to be added, but that is only with respect to
medications, and how notifications for medications will be implemented.