Update Appointment Component Appearance and Notifications
I changed the location of the save button for the appointments component
to be on the action bar. It still works the same way, but there should
be some discussion on the placement of the error message given to the
user. I also modified how notifications are created, specifically the
time in which they are set to go off. The notifications are set to go
off at 12am on the day the appointment is set to occur, unless the
appointment is created on the same day it is set to happen. If this
occurs, then the notification is set to go off an hour before. There are
probably better ways to implement this, but I think this is okay for
now.